// apps/frontend/src/lib/types/iex-cloud.ts
export interface InstitutionalOwner {
  ownerName: string
  shares: number
  positionPct?: number
  value?: number
}

export interface FundOwner {
  fundName: string
  shares: number
  positionPct?: number
  value?: number
}

export interface Ownership {
  institutionalOwnership: InstitutionalOwner[]
  fundOwnership: FundOwner[]
}